mas controller spec refactor
