+++ .g8/checkboxPage/generated-test/controllers/$className$ControllerSpec.scala
@@ -19,7 +19,6 @@ import scala.concurrent.Future
 
 class $className$ControllerSpec extends SpecBase with MockitoSugar {
 
-  def onwardRoute = Call("GET", "/foo")
 
   lazy val $className;format="decap"$Route = routes.$className$Controller.onPageLoad(NormalMode).url
