+++ .g8/multipleQuestionsPage/generated-test/controllers/$className$ControllerSpec.scala
@@ -20,7 +20,6 @@ import scala.concurrent.Future
 
 class $className$ControllerSpec extends SpecBase with MockitoSugar {
 
-  def onwardRoute = Call("GET", "/foo")
 
   val formProvider = new $className$FormProvider()
   val form = formProvider()
