+++ .g8/stringPage/generated-test/controllers/$className$ControllerSpec.scala
@@ -19,7 +19,6 @@ import scala.concurrent.Future
 
 class $className$ControllerSpec extends SpecBase with MockitoSugar {
 
-  def onwardRoute = Call("GET", "/foo")
 
   val formProvider = new $className$FormProvider()
   val form = formProvider()
