+++ test/controllers/ApplicationContactDetailsControllerSpec.scala
@@ -16,28 +16,18 @@
 
 package controllers
 
-import scala.concurrent.Future
-
-import play.api.inject.bind
-import play.api.mvc.Call
+import play.api.Application
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 
 import base.SpecBase
 import forms.ApplicationContactDetailsFormProvider
-import models.{Done, NormalMode}
-import models.ApplicationContactDetails
-import navigation.{FakeNavigator, Navigator}
-import org.mockito.ArgumentMatchers.any
-import org.mockito.Mockito.when
+import models.{ApplicationContactDetails, NormalMode}
 import org.scalatestplus.mockito.MockitoSugar
 import pages.ApplicationContactDetailsPage
-import services.UserAnswersService
 import views.html.ApplicationContactDetailsView
 
 class ApplicationContactDetailsControllerSpec extends SpecBase with MockitoSugar {
-
-  def onwardRoute = Call("GET", "/foo")
 
   val formProvider = new ApplicationContactDetailsFormProvider()
   val form         = formProvider()
@@ -99,18 +89,7 @@ class ApplicationContactDetailsControllerSpec extends SpecBase with MockitoSugar
 
     "must redirect to the next page when valid data is submitted" in {
 
-      val mockUserAnswersService = mock[UserAnswersService]
-
-      when(mockUserAnswersService.set(any())(any())) thenReturn Future.successful(Done)
-
-      val application =
-        applicationBuilder(userAnswers = Some(userAnswersAsIndividualTrader))
-          .overrides(
-            bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
-            bind[UserAnswersService].toInstance(mockUserAnswersService)
-          )
-          .build()
-
+      val application: Application = setupTestBuild(userAnswersAsIndividualTrader)
       running(application) {
         val request =
           FakeRequest(POST, applicationContactDetailsRoute)
