+++ test/controllers/HaveTheGoodsBeenSubjectToLegalChallengesControllerSpec.scala
@@ -16,27 +16,19 @@
 
 package controllers
 
-import scala.concurrent.Future
-
-import play.api.inject.bind
+import play.api.Application
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 
 import base.SpecBase
 import forms.HaveTheGoodsBeenSubjectToLegalChallengesFormProvider
-import models.{Done, NormalMode}
-import navigation.{FakeNavigator, Navigator}
-import org.mockito.ArgumentMatchers.any
-import org.mockito.Mockito.when
+import models.NormalMode
 import org.scalatestplus.mockito.MockitoSugar
 import pages.HaveTheGoodsBeenSubjectToLegalChallengesPage
-import services.UserAnswersService
 import views.html.HaveTheGoodsBeenSubjectToLegalChallengesView
 
 class HaveTheGoodsBeenSubjectToLegalChallengesControllerSpec extends SpecBase with MockitoSugar {
-
-  def onwardRoute = Call("GET", "/foo")
 
   val formProvider = new HaveTheGoodsBeenSubjectToLegalChallengesFormProvider()
   val form         = formProvider()
@@ -92,18 +84,7 @@ class HaveTheGoodsBeenSubjectToLegalChallengesControllerSpec extends SpecBase wi
 
     "must redirect to the next page when valid data is submitted" in {
 
-      val mockUserAnswersService = mock[UserAnswersService]
-
-      when(mockUserAnswersService.set(any())(any())) thenReturn Future.successful(Done)
-
-      val application =
-        applicationBuilder(userAnswers = Some(userAnswersAsIndividualTrader))
-          .overrides(
-            bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
-            bind[UserAnswersService].toInstance(mockUserAnswersService)
-          )
-          .build()
-
+      val application: Application = setupTestBuild(userAnswersAsIndividualTrader)
       running(application) {
         val request =
           FakeRequest(POST, haveTheGoodsBeenSubjectToLegalChallengesRoute)
