+++ test/controllers/SaveDraftSpec.scala
@@ -24,45 +24,44 @@ import play.api.test.FakeRequest
 import play.api.test.Helpers._
 
 import base.SpecBase
-import models.{Done, NormalMode}
+import models.Done
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.scalatestplus.mockito.MockitoSugar
-import pages.ConfidentialInformationPage
 import services.UserAnswersService
 
 trait SaveDraftSpec extends SpecBase with MockitoSugar {
-  lazy val saveDraftRoute: String =
-    routes.ExplainReasonComputedValueController.onSubmit(NormalMode, draftId, saveDraft = true).url
 
-  // Cannot pass through the route, it will error a 400
+//  lazy val saveDraftRoute: String =
+//    routes.ExplainReasonComputedValueController.onSubmit(NormalMode, draftId, saveDraft = true).url
 
-  "Redirects to Draft saved page when save-draft is selected" in {
+  def doTest(saveDraftRoute: String) =
+    "Redirects to Draft saved page when save-draft is selected" in {
 
-    val mockUserAnswersService = mock[UserAnswersService]
+      val mockUserAnswersService = mock[UserAnswersService]
 
-    when(mockUserAnswersService.set(any())(any())) thenReturn Future.successful(Done)
+      when(mockUserAnswersService.set(any())(any())) thenReturn Future.successful(Done)
 
-    val application =
-      applicationBuilder(userAnswers = Some(userAnswersAsIndividualTrader))
-        .overrides(
-          bind[UserAnswersService].toInstance(mockUserAnswersService)
-        )
-        .build()
+      val application =
+        applicationBuilder(userAnswers = Some(userAnswersAsIndividualTrader))
+          .overrides(
+            bind[UserAnswersService].toInstance(mockUserAnswersService)
+          )
+          .build()
 
-    running(application) {
-      val request =
-        FakeRequest(POST, saveDraftRoute)
-          .withFormUrlEncodedBody(("value", "answer"))
+      running(application) {
+        val request =
+          FakeRequest(POST, saveDraftRoute)
+            .withFormUrlEncodedBody(("value", "answer"))
 
-      val result = route(application, request).value
+        val result = route(application, request).value
 
-      status(result) mustEqual SEE_OTHER
-      redirectLocation(result).value mustEqual Call(
-        "POST",
-        s"/advance-valuation-ruling/$draftId/save-as-draft"
-      ).url
+        status(result) mustEqual SEE_OTHER
+        redirectLocation(result).value mustEqual Call(
+          "POST",
+          s"/advance-valuation-ruling/$draftId/save-as-draft"
+        ).url
+      }
     }
-  }
 
 }
