+++ test/controllers/ValuationMethodControllerSpec.scala
@@ -16,27 +16,18 @@
 
 package controllers
 
-import scala.concurrent.Future
-
-import play.api.inject.bind
-import play.api.mvc.Call
+import play.api.Application
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 
 import base.SpecBase
 import forms.ValuationMethodFormProvider
-import models.{Done, NormalMode, ValuationMethod}
-import navigation.{FakeNavigator, Navigator}
-import org.mockito.ArgumentMatchers.any
-import org.mockito.Mockito.when
+import models.{NormalMode, ValuationMethod}
 import org.scalatestplus.mockito.MockitoSugar
 import pages.ValuationMethodPage
-import services.UserAnswersService
 import views.html.ValuationMethodView
 
 class ValuationMethodControllerSpec extends SpecBase with MockitoSugar {
-
-  def onwardRoute = Call("GET", "/foo")
 
   lazy val valuationMethodRoute =
     routes.ValuationMethodController.onPageLoad(NormalMode, draftId).url
@@ -96,18 +87,7 @@ class ValuationMethodControllerSpec extends SpecBase with MockitoSugar {
 
     "must redirect to the next page when valid data is submitted" in {
 
-      val mockUserAnswersService = mock[UserAnswersService]
-
-      when(mockUserAnswersService.set(any())(any())) thenReturn Future.successful(Done)
-
-      val application =
-        applicationBuilder(userAnswers = Some(userAnswersAsIndividualTrader))
-          .overrides(
-            bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
-            bind[UserAnswersService].toInstance(mockUserAnswersService)
-          )
-          .build()
-
+      val application: Application = setupTestBuild(userAnswersAsIndividualTrader)
       running(application) {
         val request =
           FakeRequest(POST, valuationMethodRoute)
