+++ test/controllers/WhoAreYouAgentControllerSpec.scala
@@ -16,27 +16,18 @@
 
 package controllers
 
-import scala.concurrent.Future
-
-import play.api.inject.bind
-import play.api.mvc.Call
+import play.api.Application
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 
 import base.SpecBase
 import forms.WhoAreYouAgentFormProvider
-import models.{Done, NormalMode}
-import navigation.{FakeNavigator, Navigator}
-import org.mockito.ArgumentMatchers.any
-import org.mockito.Mockito.when
+import models.NormalMode
 import org.scalatestplus.mockito.MockitoSugar
 import pages.WhoAreYouAgentPage
-import services.UserAnswersService
 import views.html.WhoAreYouAgentView
 
 class WhoAreYouAgentControllerSpec extends SpecBase with MockitoSugar {
-
-  def onwardRoute = Call("GET", "/foo")
 
   val formProvider = new WhoAreYouAgentFormProvider()
   val form         = formProvider()
@@ -51,8 +42,6 @@ class WhoAreYouAgentControllerSpec extends SpecBase with MockitoSugar {
       val request = FakeRequest(GET, whoAreYouAgentRoute)
 
       val result = route(application, request).value
-
-      def redirectRoute = routes.AccountHomeController.onPageLoad()
 
       status(result) mustEqual OK
     }
@@ -103,18 +92,7 @@ class WhoAreYouAgentControllerSpec extends SpecBase with MockitoSugar {
 
     "must redirect to the next page when valid data is submitted" in {
 
-      val mockUserAnswersService = mock[UserAnswersService]
-
-      when(mockUserAnswersService.set(any())(any())) thenReturn Future.successful(Done)
-
-      val application =
-        applicationBuilder(userAnswers = Some(userAnswersAsIndividualTrader))
-          .overrides(
-            bind[Navigator].toInstance(new FakeNavigator(onwardRoute)),
-            bind[UserAnswersService].toInstance(mockUserAnswersService)
-          )
-          .build()
-
+      val application: Application = setupTestBuild(userAnswersAsIndividualTrader)
       running(application) {
         val request =
           FakeRequest(POST, whoAreYouAgentRoute)
